+++ cmd/systemdutil/systemdutil.go
@@ -1,9 +1,12 @@
 package systemdutil
 
 import (
+	"crypto/md5"
+	"debug/elf"
+	"encoding/hex"
 	"errors"
 	"fmt"
-	"log"
+	"io"
 	"os"
 	"os/exec"
 	"path"
@@ -16,12 +19,6 @@ import (
 	"go.aporeto.io/trireme-lib/monitor/remoteapi/client"
 	"go.aporeto.io/trireme-lib/utils/portspec"
 )
-
-var stderrlogger *log.Logger
-
-func init() {
-	stderrlogger = log.New(os.Stderr, "", 0)
-}
 
 // ExecuteCommandFromArguments processes the command from the arguments
 func ExecuteCommandFromArguments(arguments map[string]interface{}) error {
@@ -226,6 +223,9 @@ func (r *RequestProcessor) CreateAndRun(c *CLIRequest) error {
 		puType = common.HostPU
 	}
 
+	exeTags := executableTags(c)
+	c.Labels = append(c.Labels, exeTags...)
+
 	// This is added since the release_notification comes in this format
 	// Easier to massage it while creation rather than change at the receiving end depending on event
 	request := &common.EventInfo{
@@ -347,6 +347,54 @@ func sendRequest(address string, event *common.EventInfo) error {
 	}
 
 	return client.SendRequest(event)
+}
+
+func executableTags(c *CLIRequest) []string {
+
+	tags := []string{}
+
+	if fileMd5, err := computeFileMd5(c.Executable); err == nil {
+		tags = append(tags, fmt.Sprintf("@app:linux:filechecksum=%s", hex.EncodeToString(fileMd5)))
+	}
+
+	depends := libs(c.ServiceName)
+	for _, lib := range depends {
+		tags = append(tags, fmt.Sprintf("@app:linux:lib:%s=true", lib))
+	}
+
+	return tags
+}
+
+// computeFileMd5 computes the Md5 of a file
+func computeFileMd5(filePath string) ([]byte, error) {
+
+	var result []byte
+
+	file, err := os.Open(filePath)
+	if err != nil {
+		return result, err
+	}
+	defer file.Close() //nolint : errcheck
+
+	hash := md5.New()
+	if _, err := io.Copy(hash, file); err != nil {
+		return result, err
+	}
+
+	return hash.Sum(result), nil
+}
+
+// libs returns the list of dynamic library dependencies of an executable
+func libs(binpath string) []string {
+
+	f, err := elf.Open(binpath)
+	if err != nil {
+		return []string{}
+	}
+
+	libraries, _ := f.ImportedLibraries()
+	fmt.Println(libraries)
+	return libraries
 }
 
 // ParseServices parses strings with the services and returns them in an
